==> 2.Modeling/1 - integracion de datos.ipynb <==
##Liberias

import pandas as pd
import numpy as np
import warnings
warnings.filterwarnings("ignore")
import glob
import os

t_path = "C:/Users/diego/Downloads/Ciclo 2021-II/TFII/juez-bot-main/1.1.Resoluciones"

t_list = []
for year in os.listdir(t_path):
    
    c_path = t_path + "/" + year
    
    year = year[-4:]
    n_resol = len(os.listdir(c_path))
    t_list.append((year, n_resol))

for x in sorted(t_list, key=lambda x: x[0]):
    print(x)

df_cargos_desc = pd.read_excel("cargos_categorias.xlsx", 
                               sheet_name=0, 
                               usecols=[0,1,2,3])
# crear lista de los disintos cargos
cargos_list = list(df_cargos_desc["CÓDIGO"])

#df_cargos_desc.head()

df_cargos_multicat = pd.read_excel("res_hechos.xlsx", sheet_name=0, 
                                   usecols=[0, 1, 2, 3])

df_cargos_multicat["id_resol"] = \
    df_cargos_multicat["id_resol"]. \
        str.replace("–", "-"). \
        str.replace(" ", ""). \
        str.replace(".", "-"). \
        str.replace("SC2", "SPC"). \
        str.replace("TDC", "SPC")

df_cargos_multicat.info()
#df_cargos_multicat.sample(5)

#extrayendo data que proviene de 1. Cleaning-Resoluciones-Completas
df_resoluciones = pd.read_csv("res_multas.csv")

df_resoluciones["id_resol"] = \
    df_resoluciones["id_resol"]. \
        str.replace("–", "-"). \
        str.replace(" ", ""). \
        str.replace(".", "-"). \
        str.replace("SC2", "SPC"). \
        str.replace("TDC", "SPC")

df_resoluciones = df_resoluciones[["Año", "id_resol", "Empresa denunciada", "Multas", "Medidas correctivas"]]
df_resoluciones.info()
#df_resoluciones.head()

df_resol_comp = df_cargos_multicat.merge(right=df_resoluciones, how="inner", on="id_resol")
df_resol_comp = df_resol_comp.drop(["Año"], axis=1)

df_resol_comp.info()

filtro = (df_resol_comp["Multas"] == "0") & (df_resol_comp["Medidas correctivas"] == "0")

df_resol_comp["dec_fav_denunciante"] = (~filtro).astype("int8")
df_resol_comp.sample(5)

# Pre-procesamiento de las resoluciones

import re
from nltk.corpus import stopwords

def limpieza_texto(string):
    string = string.lower()
    string = string.replace("\n", " ")
    list_words = re.findall(r'[a-z0-9ñáéíóúü-]+', string)
    list_words = [x for x in list_words if x not in stopwords.words('spanish')]
    string = " ".join(list_words)
    return string

# re.findall(r'[a-z0-9ñáéíóúü-]+', "la resolución 123-a, con código asdf4.")
limpieza_texto("la resolución 123-a,\ncon código asdf4.")

df_resol_comp["cargos limpio"] = \
    df_resol_comp["cargos"].apply(limpieza_texto)

df_resol_comp["cargos limpio"]

df_resol_comp["códigos"] = \
    df_resol_comp["códigos"].apply(lambda x: x.replace(" ", "") if isinstance(x, str) else x )

df_resol_comp.to_csv("res_completo.csv", encoding="utf-8")
