==> 2.Modeling/2 - experimentos entrenamiento.ipynb <==
!pip install imbalanced-learn==0.8.1

import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt

from datetime import datetime as dt

#df_cargos_desc = pd.read_excel("cargos_categorias.xlsx", sheet_name=0)
#cargos_list = list(df_cargos_desc["CÓDIGO"])

data = pd.read_csv("res_completo.csv", index_col=0)
data.info()

# crear lista de códigos de cargos multicategoricos
data["códigos"] = data["códigos"].apply(lambda x: x.split(",") if isinstance(x, str) else [])

data.loc[:, ["id_resol", "cargos limpio", "cargos limpio stem", "dec_fav_denunciante"]]

# los datos están desbalanceados, así que tenemos que usar el precision y recall
# para evaluar sus resultados

data.groupby(["dec_fav_denunciante"])["id_resol"].count() / data["id_resol"].count()

from sklearn.pipeline import Pipeline

from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import OneHotEncoder

from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.preprocessing import MultiLabelBinarizer

from sklearn.svm import SVC
from sklearn.ensemble import RandomForestClassifier
from sklearn.neural_network import MLPClassifier
from sklearn.ensemble import GradientBoostingClassifier

from sklearn.model_selection import StratifiedKFold
from sklearn.model_selection import GridSearchCV

from imblearn.pipeline import Pipeline as PipelineIL
from imblearn.over_sampling import RandomOverSampler

# Separación del conjunto de datos en entrenamiento y test

# Crear un generador de cross-validation: 3-fold
kf = StratifiedKFold(n_splits=3, random_state=1, shuffle=True);

from sklearn.model_selection import train_test_split

# Separación del conjunto de datos en entrenamiento y test
# Separar el 20% para el test
# estratificar de acuerdo con dec_fav_denunciante

X_train, X_test, Y_train, Y_test = train_test_split(data.drop("dec_fav_denunciante", 
                                                            axis=1),
                                                    data["dec_fav_denunciante"],
                                                    test_size=0.2,
                                                    stratify=data["dec_fav_denunciante"],
                                                    random_state=0)

print("training set size: ", X_train.shape[0])
print("test set size: ", X_test.shape[0])

# Optimización de hiperparámetros mediante CV

# crear DF para almacenar todos los resultados de los experiemntos
experiments_results = pd.DataFrame(
    columns=['params', 'model_name',
             'mean_train_accuracy', 'std_train_accuracy',
             'mean_test_accuracy',  'std_test_accuracy',
             'mean_test_f1',        'std_test_f1', 
             'mean_test_precision', 'std_test_precision', 
             'mean_test_recall',    'std_test_recall'])

# función que automatiza el proceso de optimización de hiperparámetros meidante CV para un
# modelo, retorna un DF con los resultados de los experimentos

def run_CV(model, model_name, parameters, cv_iter, X, Y):
    
    clf = GridSearchCV(estimator = model,       # especificar el modelo a entrenar
                       param_grid = parameters, # especificar los parámetros a optimizar
                       cv = cv_iter,            # cross validation iterator
                        # definir métricas de evaluación
                       scoring = ['accuracy', 'precision', 'recall', 'f1'], 
                       refit = False,            # no entrenar un modelo con todos los datos
                       return_train_score=True
                      )

    # run the cv experiments
    clf.fit(X, Y)

    # store the results of the experiments in a dataframe
    keys = ['params', 'model_name',
            'mean_train_accuracy', 'std_train_accuracy',
            'mean_test_accuracy',  'std_test_accuracy',
            'mean_test_f1', 'std_test_f1', 
            'mean_test_precision', 'std_test_precision', 
            'mean_test_recall', 'std_test_recall']
             
    results_dic = {k:v for k, v in clf.cv_results_.items() if k in keys}
    results = pd.DataFrame(results_dic)
    results["model_name"] = model_name
    
    return results

def report_metrics(data):

    fig, axis = plt.subplots( 1, 3, figsize=(17, 5))

    sns.scatterplot(x="mean_test_precision", 
                  y="mean_test_recall",
                  data=data,
                  ax=axis[0])

    sns.scatterplot(x="mean_test_accuracy", 
                  y="mean_test_f1",
                  data=data,
                  ax=axis[1])

    sns.scatterplot(x="mean_train_accuracy", 
                  y="mean_test_accuracy",
                  data=data,
                  ax=axis[2])

    axis[0].set_ylim(0.0, 1.0)
    axis[1].set_ylim(0.0, 1.0)
    axis[2].set_ylim(0.0, 1.0)

    axis[0].set_xlim(0.0, 1.0)
    axis[1].set_xlim(0.0, 1.0)
    axis[2].set_xlim(0.0, 1.0)

    plt.show()


## model config

columnas = ["cargos limpio"]

# # Original arguments
# default_ngrams_args = {
#     'FeaExt__max_df': [1.0, 0.9], 
#     'FeaExt__min_df': [0.0, 0.1],
#     'FeaExt__ngram_range': [(1, 1), (2, 2), (1, 2)]
# }

# default_tfidf_args = {
#     'FeaExt__max_df': [1.0, 0.9], 
#     'FeaExt__min_df': [0.0, 0.1],
#     'FeaExt__ngram_range': [(1, 1), (2, 2), (1, 2)]
# }

# default_mlp_args = {
#     'Pred__random_state': [0],
#     'Pred__max_iter': [200],
#     'Pred__tol': [0.001],
#     'Pred__learning_rate_init': [0.001],
#     'Pred__activation': ['tanh', 'relu'],
#     #'Pred__alpha': [0.0001, 0.001],
#     'Pred__batch_size': [25, 'auto'],
#     'Pred__hidden_layer_sizes': [(10), (10, 5)],
# }

# default_rf_args = {
#   'Pred__random_state': [0], 
#   'Pred__criterion': ["entropy", "giny"], 
#   'Pred__n_estimators': [100],
#   'Pred__min_samples_leaf':[3, 5, 7],
#   'Pred__bootstrap': [True, False],
#   'Pred__max_samples': [0.95, 0.90],
# }

# default_svm_args ={
#     'Pred__random_state': [0], 
#     'Pred__C': [0.1, 1, 10], 
#     'Pred__kernel': ['linear', 'rbf'],
# }

default_ngrams_args = {
    'FeaExt__max_df': [1.0, 0.9], 
    'FeaExt__min_df': [0.0, 0.1],
    'FeaExt__ngram_range': [(1, 1), (2, 2), (1, 2)]
}

default_tfidf_args = {
    'FeaExt__max_df': [1.0, 0.9], 
    'FeaExt__min_df': [0.0, 0.1],
    'FeaExt__ngram_range': [(1, 1), (2, 2), (1, 2)]
}

default_mlp_args = {
    'Pred__random_state': [0],
    'Pred__max_iter': [200],
    'Pred__tol': [0.001],
    'Pred__learning_rate_init': [0.001],
    'Pred__activation': ['tanh', 'relu'],
    'Pred__batch_size': [25, 'auto'],
    'Pred__hidden_layer_sizes': [(10), (10, 5)],
}

default_rf_args = {
  'Pred__random_state': [0], 
  'Pred__criterion': ["entropy",], 
  'Pred__n_estimators': [100],
  'Pred__min_samples_leaf':[1, 3, 5, 7],
  'Pred__bootstrap': [True, False],
  'Pred__max_samples': [0.95, 0.90],
}

default_svm_args ={
    'Pred__random_state': [0], 
    'Pred__C': [0.1, 1, 10], 
    'Pred__kernel': ['linear', 'rbf'],
}

modelos = [
    
    # N-GRAMAS + MLP
    {"model": "ngramas + mlp", 
     "pipeline": Pipeline([("FeaExt", CountVectorizer()), ('Pred', MLPClassifier())]),
     "params": {**default_ngrams_args, 
                **default_mlp_args}
    },
    

    # N-GRAMAS + RF
     {"model": "ngramas + rf", 
      "pipeline": Pipeline([("FeaExt", CountVectorizer()), ('Pred', RandomForestClassifier())]),
      "params": {
                **default_ngrams_args,
                **default_rf_args,
                'Pred__class_weight': ['balanced', None]}
     },


    # N-GRAMAS + SVM
     {"model": "ngramas + svm", 
      "pipeline": Pipeline([("FeaExt", CountVectorizer()), ('Pred', SVC())]),
      "params": {
                **default_ngrams_args,
                **default_svm_args,
                'Pred__class_weight': [None, 'balanced']
                 }
     },



    # OVERSAMPLING + N-GRAMAS + MLP
     {"model": "oversampling + ngramas + mlp", 
      "pipeline": PipelineIL([
                  # when predicting, this step does nothing
                  # resamples each CV fold and not the entire training set 
                  ("FeaExt", CountVectorizer()),
                  ("Oversampling", RandomOverSampler(random_state=0, sampling_strategy='not majority')),
                  ('Pred', MLPClassifier())
                  ]),
      "params": {
                **default_ngrams_args, 
                **default_mlp_args
                }
      },


    # OVERSAMPLING + N-GRAMAS + RF
     {"model": "oversampling + ngramas + rf", 
      "pipeline": PipelineIL([ 
                  ("FeaExt", CountVectorizer()),
                  ("Oversampling", RandomOverSampler(random_state=0, sampling_strategy='not majority')),
                  ('Pred', RandomForestClassifier())
                  ]),
      "params": {
                **default_ngrams_args,
                **default_rf_args,
                 }
      },


    # OVERSAMPLING + N-GRAMAS + SVM
     {"model": "oversampling + ngramas + svm", 
      "pipeline": PipelineIL([ 
                  ("FeaExt", CountVectorizer()),
                  ("Oversampling", RandomOverSampler(random_state=0, sampling_strategy='not majority')),
                  ('Pred', SVC())
                  ]),
      "params": {
                **default_ngrams_args,
                **default_svm_args,
                 }
      },


    # TF-IDF + MLP
    {"model": "tf-idf + mlp", 
     "pipeline": Pipeline([("FeaExt", TfidfVectorizer()), ('Pred', MLPClassifier())]),
     "params": {
                **default_tfidf_args, 
                **default_mlp_args
                }
    },


    # TF-IDF + RF
    {"model": "tf-idf + rf", 
     "pipeline": Pipeline([("FeaExt", TfidfVectorizer()), ('Pred', RandomForestClassifier())]),
     "params": {
                **default_tfidf_args,
                **default_rf_args,
                'Pred__class_weight': ['balanced', None]}
    },


    # TF-IDF + SVM
    {"model": "tf-idf + svm", 
     "pipeline": Pipeline([("FeaExt", TfidfVectorizer()), ('Pred', SVC())]),
     "params": {
                **default_tfidf_args,
                **default_svm_args,
                'Pred__class_weight': [None, 'balanced']
                }
    },

]

for columna in columnas:
    for model_spec in modelos:

        pipeline = model_spec["pipeline"]
        params =  model_spec["params"]
        name = model_spec["model"]

        print(f"entrenando modelo: {name} - {columna}")
        
        # seleccionar la columna
        X_train_temp = X_train.loc[:, [columna]].values.reshape((X_train.shape[0]))
        
        # CV
        exp_results_df = run_CV(pipeline, 
                                f"{name} - {columna}", 
                                params, 
                                kf, 
                                X_train_temp, Y_train)
        
        # almacenar los resultados
        experiments_results = experiments_results.append(exp_results_df)

    # guardar los resultados
    time_sv = dt.now().strftime("%m-%d-%H-%M")
    experiments_results.to_csv(f"resultados_{time_sv}.csv")

experiments_results
