==> 2.Modeling/3 - entrenamiento y test.ipynb <==
import pandas as pd
import seaborn as sns
import numpy as np
import os
import re

from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import OneHotEncoder

from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.preprocessing import MultiLabelBinarizer

from sklearn.svm import SVC
from sklearn.ensemble import RandomForestClassifier
from sklearn.neural_network import MLPClassifier
from sklearn.ensemble import GradientBoostingClassifier

from sklearn.model_selection import KFold, StratifiedKFold
from sklearn.model_selection import GridSearchCV, cross_val_score, train_test_split

from sklearn.metrics import accuracy_score, f1_score, precision_score, recall_score

from imblearn.pipeline import Pipeline as PipelineIL
from imblearn.over_sampling import RandomOverSampler

from sklearn.pipeline import Pipeline

# leer todos los archivos de resultados

df_resultados = pd.DataFrame()
for file in os.listdir("resultados"):
    temp_df = pd.read_csv(f"resultados/{file}",  index_col=0)
    df_resultados = df_resultados.append(temp_df)

df_resultados

def is_balanced(string):
    result = "'Pred__class_weight': 'balanced'" in string
    if result:
        return "balanced"
    else:
        return "not balanced"

df_resultados["model_name"] = df_resultados["model_name"] + " - " + df_resultados["params"].apply(is_balanced)

df_resultados

# para cada modelo
# identificar el mayor cv_f1
# identificar los hiperparámetros que generaron ese score

top_result_by_model = pd.DataFrame(columns=df_resultados.columns)

for model_name in df_resultados["model_name"].unique():
    model_results = df_resultados[df_resultados["model_name"] == model_name]
    max_test_f1 = model_results["mean_test_f1"].max()
    
    best_model_results = model_results[model_results["mean_test_f1"] == max_test_f1].head(1)
    top_result_by_model = top_result_by_model.append(best_model_results)


print(len(top_result_by_model))
top_result_by_model.sort_values(by="mean_test_accuracy", ascending=False)

# extraer características adicionales de los modelos

def get_model(string):
    string, _, _ = string.split(" - ")
    string = string.split(" + ")[-1]
    return string

top_result_by_model["oversampled"] = top_result_by_model["model_name"].str.contains("oversampling")
top_result_by_model["ngramas"] = top_result_by_model["model_name"].str.contains("ngramas")
top_result_by_model["class_weights"] = top_result_by_model["params"].str.contains("balanced")

top_result_by_model["modelo"] = top_result_by_model["model_name"].apply(get_model)

top_result_by_model

top_result_by_model.iloc[1, 0]

# leer el conjunto de datos

data = pd.read_csv("res_completo.csv", index_col=0)
data.info()

# Separación del conjunto de datos en entrenamiento y test
# Separar el 20% para el test
# estratificar de acuerdo con dec_fav_denunciante

X_train, X_test, Y_train, Y_test = train_test_split(data.drop("dec_fav_denunciante", 
                                                            axis=1),
                                                    data["dec_fav_denunciante"],
                                                    test_size=0.2,
                                                    stratify=data["dec_fav_denunciante"],
                                                    random_state=0)

print("training set size: ", X_train.shape[0])
print("test set size: ", X_test.shape[0])

# Crear un generador de cross-validation: 3-fold
kf = StratifiedKFold(n_splits=3, random_state=1, shuffle=True);

def args_string_to_dict(arguments):
    arguments = arguments[1:-1]
    arguments = arguments.split(", '")
    arguments = [ x.replace("'", "") for x in arguments ]
    arguments = [ tuple(x.split(": ")) for x in arguments ]

    def eval_obj(string):
        if re.match("[a-zA-Z]+", string) and string != "None":
            if string == "None":
                return None
            elif string == "True":
                return True
            elif string == "False":
                return False
            else:
                return string
        else: 
            return eval(string)

    arguments = [ (x, eval_obj(y)) for x, y in arguments]
    arguments = {k:v for k,v in arguments}

    return arguments

def separate_args(args_dict):
    # extract FE and model arguments to pass to the 
    fe_args = {k: v for k,v in args_dict.items() if k.startswith("FeaExt")}
    fe_args = {k.split("__")[-1]: v for k,v in fe_args.items()}

    model_args = {k: v for k,v in args_dict.items() if k.startswith("Pred")}
    model_args = {k.split("__")[-1]: v for k,v in model_args.items()}
    
    return fe_args, model_args

def get_best_arg(arg_desc_string):
    #arg_desc_string = top_result_by_model[top_result_by_model["model_name"]==model].iloc[indx,0]
    dict_rgs = args_string_to_dict(arg_desc_string)
    fe_args, model_args = separate_args(dict_rgs)
    
    return fe_args, model_args

def create_pipeline(pipeline_descr, fe_args={}, pred_args={}):
    # dada la descripción de un modelo, reconstruir el pipeline

    model_desc, variable, _ = pipeline_descr.split(" - ")
    steps = model_desc.split(" + ")

    pipeline_steps = []

    if "ngramas" in steps:
        pipeline_steps.append(("FeaExt", CountVectorizer(**fe_args)))
    if "tf-idf" in steps:
        pipeline_steps.append(("FeaExt", TfidfVectorizer(**fe_args)))
    
    
    if "oversampling" in steps:
        pipeline_steps.append(("Oversampling", RandomOverSampler(random_state=0, sampling_strategy='not majority')))

    
    if "mlp" in steps:
        pipeline_steps.append(('Pred', MLPClassifier(**pred_args)))
    if "rf" in steps:
        pipeline_steps.append(('Pred', RandomForestClassifier(**pred_args)))
    if "svm" in steps:
        pipeline_steps.append(('Pred', SVC(**pred_args)))                  
    
    pipeline = PipelineIL(pipeline_steps)
    
    return pipeline

def evaluate_model_on_test(row, debug=False):
    # extraer las características del modelo
    model_name = row["model_name"]
    #print(f"model desc: {model_name}")
    
    # extraer la columna usada en la predicción
    _, columna, _ = row["model_name"].split(" - ")
    #print(f"input col: {columna}")

    # extraer los hiperparametros utilizados para entrenar el modelo
    fe_args, model_args = get_best_arg(row["params"])
    #print(fe_args)
    #print(model_args)

    # crear el pipeline
    pipeline = create_pipeline(model_name, fe_args, model_args)
    #print(pipeline)
    
    if debug:
        print(row["params"])
        print()
        print(model_name)
        print(columna)
        print(fe_args)
        print(model_args)
        print(pipeline)
        return None

    # entrenar el modelo
    X_train_temp = X_train.loc[:, [columna]].values.reshape((X_train.shape[0]))
    X_test_temp = X_test.loc[:, [columna]].values.reshape((X_test.shape[0]))
    pipeline.fit(X_train_temp, Y_train)
    
    # evaluar el modelo en test
    #print("test predictions")
    # keys: ['0', '1', 'accuracy', 'macro avg', 'weighted avg']    
    Y_test_pred = pipeline.predict(X_test_temp)
    
    acc_test_score = accuracy_score(y_true = Y_test, y_pred = Y_test_pred)
    f1_test_score = f1_score(y_true = Y_test, y_pred = Y_test_pred)
    prec_test_score = precision_score(y_true = Y_test, y_pred = Y_test_pred)
    rec_test_score = recall_score(y_true = Y_test, y_pred = Y_test_pred)


    return acc_test_score, f1_test_score, prec_test_score, rec_test_score

# testeo de la función

for indx, row in top_result_by_model.sample(1).iterrows():
    print(evaluate_model_on_test(row, debug=False))

# para cada row en top_result_by_model reconstruir el modelo, entrenarlo y evaluarlo en el test
top_result_by_model["test_metrics"] =  top_result_by_model.apply(evaluate_model_on_test, axis=1)

# extraer las métricas de evaluación en el conjunto de datos de test
top_result_by_model["test_accuracy"] = top_result_by_model["test_metrics"].apply(lambda x: x[0])
top_result_by_model["test_f1"] = top_result_by_model["test_metrics"].apply(lambda x: x[1])
top_result_by_model["test_precision"] = top_result_by_model["test_metrics"].apply(lambda x: x[2])
top_result_by_model["test_recall"] = top_result_by_model["test_metrics"].apply(lambda x: x[3])

top_result_by_model

top_result_by_model = top_result_by_model.rename(
    columns={'mean_test_f1': 'CV F1',
             'mean_test_precision': 'CV Prec',
             'mean_test_recall': 'CV Rec',
             'mean_test_accuracy': 'CV Acc',
             'model_name': 'Modelo',
             'params': 'Hiper parámetros'
            })

top_result_by_model

# guardar los resultados en una tabla

top_result_by_model.to_csv("mejores_modelos/resultados_definitivos.csv")

## Tabla de evaluación de modelos finales


EXPERIMENTOS
- ngramas vs ngramas + td-idf
- no oversampling vs oversampling

df = top_result_by_model

model_names_dic = {"mlp": "Redes Neuronales", "svm": "SVM", "rf": "Random Forest"}
df["modelo"] = df["modelo"].map(model_names_dic)

# Experimentos 2: resultados en entrenamiento

temp_df = df[df["ngramas"]==True]

temp_df = temp_df.loc[:, ["CV Acc", "CV F1", "CV Prec", "CV Rec", "ngramas", "modelo", "oversampled", "class_weights"]]

def model_type(row):
    if row["oversampled"]:
        return "oversampled"
    elif row["class_weights"]:
        return "weighted"
    else:
        return "none"

temp_df["model_type"] = temp_df.apply(model_type, axis=1)

temp_df = temp_df.drop(["ngramas", "oversampled", "class_weights"], axis=1)

temp_df = temp_df.melt(id_vars=['model_type', "modelo"], var_name='metric', value_name='value')

temp_df["model_type"] = temp_df["model_type"].map(
    {"none": "Ninguno", 
     "weighted": "Pesos de Clase", 
     "oversampled": "Sobremuestreo"})

temp_df = temp_df.rename(columns={"metric": "Métrica", "model_type": "Tipo", "modelo": "Modelo"})

temp_df = pd.pivot_table(temp_df, 
                         values='value', 
                         index=['Tipo'],
                         columns=['Modelo', 'Métrica'],)

#temp_df
for var in ["Random Forest", "Redes Neuronales", "SVM"]:
    print(r"""
\begin{table}[h!]
\centering"""
         )
    print(temp_df.loc[:, [var]].round(3).to_latex(na_rep="-"))
    print("""
\\caption{Resultados del modelo \\textit{Random Forest} en la etapa de entrenamiento}
\\label{Exp 1 - entrenamiento - rn}
\\end{table}
    """)
    print()

# Experimentos 2: resultados en prueba

temp_df = df[df["ngramas"]==True]

temp_df = temp_df.loc[:, ["test_accuracy", "test_f1", "test_precision", "test_recall", "ngramas", "modelo", "oversampled", "class_weights"]]

def model_type(row):
    if row["oversampled"]:
        return "oversampled"
    elif row["class_weights"]:
        return "weighted"
    else:
        return "none"

temp_df["model_type"] = temp_df.apply(model_type, axis=1)

temp_df = temp_df.drop(["ngramas", "oversampled", "class_weights"], axis=1)

temp_df = temp_df.melt(id_vars=['model_type', "modelo"], var_name='metric', value_name='value')

temp_dict = {"test_accuracy": "Acc Prueba", 
             "test_f1": "F1 Prueba", 
             "test_recall": "Rec Prueba",
             "test_precision": "Prec Prueba"}

temp_df["metric"] = temp_df["metric"].map(temp_dict)

temp_df["model_type"] = temp_df["model_type"].map(
    {"none": "Ninguno", 
     "weighted": "Pesos de Clase", 
     "oversampled": "Sobremuestreo"})

temp_df = temp_df.rename(columns={"metric": "Métrica", "model_type": "Tipo", "modelo": "Modelo"})


temp_df = pd.pivot_table(temp_df, 
                         values='value', 
                         index=['Tipo'],
                         columns=['Modelo', 'Métrica'],)

#temp_df.round(3)

for var in ["Random Forest", "Redes Neuronales", "SVM"]:
    print(temp_df.loc[:, [var]].round(3).to_latex(na_rep="-"))
    print()

# Experimentos 1: resultados en entrenamiento

temp_df = df[(df["oversampled"]==False) & (df["class_weights"]==False)]

temp_df = temp_df.loc[:, ["CV Acc", "CV F1", "CV Prec", "CV Rec", "ngramas", "modelo"]]

temp_df["ngramas"] = temp_df["ngramas"].apply(lambda x: "n-gramas" if x else "tf-idf")

temp_df = temp_df.melt(id_vars=['ngramas', 'modelo'], var_name='metric', value_name='value')

temp_df = temp_df.rename(columns={"metric": "Métrica", "ngramas": "Tipo", "modelo": "Modelo"})

temp_df = pd.pivot_table(temp_df, 
                         values='value', 
                         index=['Tipo'],
                         columns=['Modelo', 'Métrica'])

#temp_df.round(3)
for var in ["Random Forest", "Redes Neuronales", "SVM"]:
    print(temp_df.loc[:, [var]].round(3).to_latex(na_rep="-"))
    print()

# Experimentos 1: resultados en prueba

temp_df = df[(df["oversampled"]==False) & (df["class_weights"]==False)]

temp_df = temp_df.loc[:, ["test_accuracy", "test_f1", "test_precision", "test_recall", "ngramas", "modelo"]]

temp_df["ngramas"] = temp_df["ngramas"].apply(lambda x: "n-gramas" if x else "tf-idf")

temp_df = temp_df.melt(id_vars=['ngramas', 'modelo'], var_name='metric', value_name='value')

temp_dict = {"test_accuracy": "Acc Prueba", 
             "test_f1": "F1 Prueba", 
             "test_recall": "Rec Prueba",
             "test_precision": "Prec Prueba"}

temp_df["metric"] = temp_df["metric"].map(temp_dict)

temp_df = temp_df.rename(columns={"metric": "Métrica", "ngramas": "Tipo", "modelo": "Modelo"})

temp_df = pd.pivot_table(temp_df, 
                         values='value', 
                         index=['Tipo'],
                         columns=['Modelo', 'Métrica'])

#temp_df.round(3)

for var in ["Random Forest", "Redes Neuronales", "SVM"]:
    print(temp_df.loc[:, [var]].round(3).to_latex(na_rep="-"))
    print()

# Comparación n-gramas vs t-idf: No hay diferencia aparentemente
df[df["oversampled"]==False].loc[:, ["model_name", "test_accuracy", "test_f1", "test_precision", "test_recall"]]

# Comparación no oevrsampling vs oversampling
df[df["ngramas"]==True].loc[:, ["model_name", "test_accuracy", "test_f1", "test_precision", "test_recall"]]
